requirements.txt
# 核心框架
fastapi==0.115.0
uvicorn[standard]==0.30.1
streamlit==1.36.0
python-multipart==0.0.9

# 通义千问 API
openai>=1.58.1, <2.0.0  # 让 pip 自动选择兼容版本
dashscope==1.20.1

# LangChain 生态
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langsmith>=0.1.125,<0.3
chromadb==0.5.5  # 替换 faiss-cpu

# 文档解析
pypdf2==3.0.1
python-docx==1.1.0
unstructured==0.12.0

# 数据库
sqlalchemy==2.0.23
sqlite-utils==3.36

# 数据处理
numpy
pandas
plotly==5.18.0

# Bot 框架
python-telegram-bot==21.0.1
botbuilder-core==4.14.7
aiohttp==3.9.1

# 工具
requests==2.32.0
pydantic==2.9.0
python-dotenv==1.0.0